add_data property added. Test were moved to another directory. Time
