+++ ipysurfer/__init__.py
@@ -1,3 +1,4 @@
 """Google Summer of Code 3D rendering using ipyvolume."""
 
 from .viz import Brain
+from .stc import plot_source_estimates
