+++ ipysurfer/_utils/__init__.py
@@ -1,5 +1,5 @@
 """Utility functions and classes."""
 
+from .colormap import _calculate_cmap
 from .funcs import (_check_units, _compute_normals,
                     _fast_cross_3d, _get_subjects_dir)
-from .surface import Surface
