+++ ipysurfer/tests/conftest.py
@@ -8,7 +8,7 @@ import pytest
 SubjectInfo = namedtuple('SubjectInfo', 'subject_id subjects_dir')
 
 
-@pytest.fixture(scope='module')
+@pytest.fixture(scope='session')
 def get_subject_info():
     data_path = mne.datasets.sample.data_path()
     subjects_dir = path.join(data_path, 'subjects')
