+++ ipysurfer/viz/__init__.py
@@ -1,0 +1,4 @@
+"""Visualization related tools."""
+
+from .brain import Brain
+from .surface import Surface
