+++ ipysurfer/viz/time_viewer.py
@@ -1,0 +1,4 @@
+class TimeViewer:
+    """Stab class for time viewer."""
+
+    pass
